// server/src/datasets/congress-legislators-types.ts
/**
 * Types for [congress-legislators](https://github.com/unitedstates/congress-legislators)
 * dataset
 *
 * This is autogenerated using quicktype.io
 */
/** */
export interface Legislator {
  id: ID;
  name: Name;
  bio: Bio;
  terms: Term[];
  leadership_roles?: LeadershipRole[];
  other_names?: OtherName[];
  family?: Family[];
}

export interface Bio {
  birthday: string;
  gender: Gender;
}

export enum Gender {
  F = "F",
  M = "M"
}

export interface Family {
  name: string;
  relation: string;
}

export interface ID {
  bioguide: string;
  thomas?: string;
  lis?: string;
  govtrack: number;
  opensecrets: string;
  votesmart?: number;
  fec: string[];
  cspan?: number;
  wikipedia: string;
  house_history?: number;
  ballotpedia?: string;
  maplight?: number;
  icpsr?: number;
  wikidata: string;
  google_entity_id?: string;
}

export interface LeadershipRole {
  title: string;
  chamber: Chamber;
  start: string;
  end?: string;
}

export enum Chamber {
  House = "house",
  Senate = "senate"
}

export interface Name {
  first: string;
  last: string;
  official_full: string;
  middle?: string;
  nickname?: string;
  suffix?: Suffix;
}

export enum Suffix {
  Ii = "II",
  Iii = "III",
  Iv = "IV",
  Jr = "Jr."
}

export interface OtherName {
  last: string;
}

export interface Term {
  type: Type;
  start: string;
  end: string;
  state: string;
  district?: number;
  party: Caucus;
  url?: string;
  class?: number;
  address?: string;
  phone?: string;
  fax?: null | string;
  contact_form?: string;
  office?: string;
  state_rank?: StateRank;
  rss_url?: string;
  how?: How;
  caucus?: Caucus;
  party_affiliations?: PartyAffiliation[];
  "end-type"?: string;
}

export enum Caucus {
  Democrat = "Democrat",
  Independent = "Independent",
  Republican = "Republican"
}

export enum How {
  Appointment = "appointment"
}

export interface PartyAffiliation {
  start: string;
  end: string;
  party: Caucus;
}

export enum StateRank {
  Junior = "junior",
  Senior = "senior"
}

export enum Type {
  Rep = "rep",
  Sen = "sen"
}